Upgrade boost
Former-commit-id: fdcf8c32898bc60524228c3492ef1a3c3a98c51f

+++ dependencies64/boost/boost/wave/util/cpp_macromap.hpp
@@ -742,7 +742,7 @@ token_type startof_argument_list = *next;
             return 0;
         }
 
-        switch (static_cast<unsigned int>(id)) {
+        switch (id) {
         case T_LEFTPAREN:
             ++nested_parenthesis_level;
             argument->push_back(*next);
@@ -1414,7 +1414,7 @@ string_type const &value = curr_token.get_value();
     char buffer[22];    // 21 bytes holds all NUL-terminated unsigned 64-bit numbers
 
         using namespace std;    // for some systems sprintf is in namespace std
-        sprintf(buffer, "%ld", main_pos.get_line());
+        sprintf(buffer, "%zd", main_pos.get_line());
         expanded.push_back(token_type(T_INTLIT, buffer, curr_token.get_position()));
         return true;
     }
